+++ src/components/custom/sortable-list/index.ts
@@ -1,4 +1,8 @@
 export { default as SortableList } from "./SortableList.vue";
 export { default as SortableItem } from "./SortableItem.vue";
 export { default as SortableHandle } from "./SortableHandle.vue";
-export { SORTABLE_CONTEXT } from "./context";
+export {
+    SORTABLE_CONTEXT,
+    provideSortableContext,
+    useSortableContext,
+} from "./context";